convert to Issue and add issue support
